projects/chatbot-server/requirements.txt
django
django-grappelli
django-redis

djangorestframework
markdown
django-filter
